+++ changelog.md
@@ -6,5 +6,7 @@ Adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [ Unreleased(v0.2.3) ]
 
 ### Added
+- Added help printing.
+- Added 'status' subcommand.
 - Added 'update' subcommand.
 - Added 'download' subcommand.